Refactor project structure and update namespaces
Moved data files and DBFacade to appropriate infrastructure and scripts directories. Updated namespaces across domain, application, and infrastructure layers for consistency. Added missing project references and Microsoft.Data.Sqlite package to Chirp.Infrastructure. Refactored service and page dependencies to use new interfaces and namespaces.
