+++ src/Chirp.Infrastructure/Chirp.Infrastructure.csproj
@@ -5,6 +5,10 @@
     <ProjectReference Include="..\Chirp.Application\Chirp.Application.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.9" />
+  </ItemGroup>
+
   <PropertyGroup>
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
