+++ src/Chirp.Infrastructure/Services/CheepService.cs
@@ -1,4 +1,9 @@
-namespace Chirp.Razor.Data;
+using Chirp.Domain.Entities;
+using Chirp.Application.Interfaces;
+using Chirp.Infrastructure.Data;
+
+
+namespace Chirp.Infrastructure.Services;
 public class CheepService : ICheepService
 {
     private readonly DBFacade _dbFacade;
