+++ src/Chirp.Razor/Chirp.Razor.csproj
@@ -18,4 +18,10 @@
     <PackageReference Include="Microsoft.Data.Sqlite" Version="9.0.9" />
   </ItemGroup>
 
+
+  <ItemGroup>
+    <ProjectReference Include="..\Chirp.Application\Chirp.Application.csproj" />
+    <ProjectReference Include="..\Chirp.Infrastructure\Chirp.Infrastructure.csproj" />
+  </ItemGroup>
+
 </Project>
