+++ src/Chirp.Razor/Pages/UserTimeline.cshtml
@@ -1,5 +1,4 @@
 ﻿@page "/{author}"
-@using Chirp.Razor.Data
 @model Chirp.Razor.Pages.UserTimelineModel
 @{
     ViewData["Title"] = "Chirp!";
